Video-Image_Analysis/requirements.txt
streamlit
opencv-python
google-cloud-vision
python-dotenv
google-cloud-vision
opencv